refactor(core): extract consts tests to sibling file
primer-core/src/consts.rs carried a single inline trailing test module
(#[cfg(test)] mod tests @561) of consts-sanity assertions. Move it
verbatim into a sibling src/consts/tests.rs, replacing the inline block
with `#[cfg(test)] mod tests;`.

603 -> 562 lines (-41), now under the ~500 guideline's neighbourhood and
consistent with the 16 prior test-module extractions (#268-#283). The
test module is a descendant of `consts`, so every `use super::*` /
`use super::inference::*` resolves verbatim -- no visibility edits.

Test-only relocation; no behaviour change. primer-core --lib 176 passed
(identical to baseline); workspace 51 ok blocks, 0 failures; clippy and
fmt clean.
